Add recursion limit to Should -Be when comparing arrays (#824)
Should Be unfolds arrays to compare their internal values. Providing cyclic array to ArraysAreEqual function would cause the code to overflow. We impose a limit of 100 layers of array (that is array that contains another array, that contains another array, and so on 100 times), because we want the assertion to fail fast, and we asssume that 100-layer array is not something user would create on purpose. 

Fix #785