Fix tags filter warning
Fixes tag warning that appears when we run directly from an editor because the test stack in that case is smaller than 2 and we were only looking at the stack size to determine how deep we are in the structure. This fix adds a flag to the Pester state that determines if we are running directly or via Invoke-Pester by using a switch when we instantiate state directly in Invoke-Pester and no switch when we instantiate in Describe.