Add space between 'Examples:' and examples name
When enumerating the Gherkin scenarios, if there are scenario example
tables, the scenario name is augmented to append a newline and then the
text 'Examples:' followed by the name of the example table (if
provided). This change, instead, appends 'Examples: ' to the scenario
name to provide a space between 'Examples:' and the name of the examples
table.