Correctly report number of found step files
Fixes #1149

It appears that at one point, there was a variable named $StepFiles and
it was used to report how many step files from which step definitions
were loaded when displaying Verbose output. At some later time,
$StepFiles was removed and inlined into the foreach loop.

This commit restores the $StepFiles intermediate variable so that the
proper verbose output is displayed.