Added possibility to pass tests script as string to address cases whe… (#972)
Added possibility to pass tests script as string to address cases when powershell scripts and tests for them are stored not on FileSystem but in some storage like DataBase and executed in Temp folder.
Example is Sitecore CMS, where we use SPE (Sitecore Powershell Extensions) module to create and run PS scripts. We store scripts and tests as Items in content tree and SPE executes them when needed in Temp folder on the server.