Check scope of Set-Variable to resolve #1107 (#1109)
Calling through mock to Set-Variable adds two scopes above it, so we need to patch the call to do the correct thing. When no scope, local or 0 scope is used we use 2, when number scope is defined we use scope number + 2, and if script or global is used we do nothing because those scopes are correctly.