Fix features and scenario names do not display when PesterResult is printed to the console (#1138)
* "Override" ToString on Gherkin.Ast.Feature

This commit adds a types.ps1xml file to add an "override" of the
ToString method as a ScriptMethod to Gherkin.Ast.Feature objects. This
will allow the feature names to be displayed when using -PassThru with
Invoke-Gherkin and printing the PesterResult object to the console.

* Add space between 'Examples:' and examples name

When enumerating the Gherkin scenarios, if there are scenario example
tables, the scenario name is augmented to append a newline and then the
text 'Examples:' followed by the name of the example table (if
provided). This change, instead, appends 'Examples: ' to the scenario
name to provide a space between 'Examples:' and the name of the examples
table.

* "Override" ToString on Gherkin.Ast.Scenario

This commit updates the Gherkin.types.ps1xml file to add an "override"
of the ToString method as a ScriptMethod to Gherkin.Ast.Scenario
objects. This will allow the scenario names to be displayed when using
-PassThru with Invoke-Gherkin and printing the PesterResult object to
the console.