Session state hints (#1169)
Add debugging output that captures all session states and describes them by which module they come from or where we captured them. Based on the output it is then easy to see in which scope the code is invoked and makes debugging problems easy. It also shows how deep in the session state we are running. So given a scripblock $sb = {} invoking it like this &$sb would report 0 scopes deep, and &{ &{ &{ &$sb } } } would report 3 scopes deep. The output is disabled by default, and can be enabled by providing @{ Experimental = $true; ShowScopeHints = $true } Pester option to Invoke-Pester.