Ensure $null is on LHS of equality checks
PSScriptAnalyzer warns that $null should be on the LHS of comparison
operators to prevent subtle comparison bugs from creeping in. Yup, that
most certainly is true.

Except for one case in Coverage.ps1. Which is exactly what
PSScriptAnalyzer is warning about--but in this case, is the desired
behavior of the code.

There is (to me, anyway) an obscure fact about PowerShell comparison
operators: if the "input" (i.e. the LHS) to the operator is a collection
object (such as an array), then the comparison operator acts as a filter
and compares each item in the collection with the RHS operand and
returns a new collection containing the items which satisfy the
predicate/filter.

I added a code comment above this line so that others (such as myself),
who may not be as well versed about PowerShell will know not to change
this.

This commit also contains some whitespace changes and fixes to
capitalization.