Assert-MockCalled -Scope: allow Feature and Scenario (#1170)
To support asserting mocks with Assert-MockCalled for Gherkin-style
tests, allow the Scope parameter to also accept the "keywords" Feature
and Scenario.

Prior to this fix, using the Scope parameter with anything but a numeric
scope was not possible. Now you can specify that the scope associated
with the Gherkin Feature or Scenario should be used when asserting a
mock.