Don't use the foreach alias
It's a best practice to not use aliases when writing modules because you
can't be sure someone didn't remove or change the alias--stranger things
have happened. It so happens, the calls to foreach which I updated were
not making use of $SafeCommands--and for the life of me, I still don't
fully understand when one should be using the $SafeCommand instance of
the command versus just outright calling the command directly.