@wip Gherkin v2 (or should I say, v5?)
OK, so this is the very first beginnings of revamping the Gherkin side
of Pester. So, don't be too critical.

I want to be able to benefit from the wisdom of others, i.e. the
original implementation, so I didn't want to go and wholesale start
changing Invoke-Gherkin proper (you'll already see lots of
customizations I made when I thought I could "shoehorn" in the new
functionality to the existing framework). So instead, I've opted to
create a Gherkin2.ps1 file. This is TEMPORARY until I at least get a
modicum of functionality on par with what's already present in Pester
v4.

Also, I'm not happy with the structure of the tests, but these are early
days and I foresee the structure changing greatly as work progresses.

Anyway, I'm throwing this out there to get the juices flowing! Oh, and I
fully expect the tests not to be passing with this commit at this time.