@wip Adding Gherkin3.ps1 and tests
So, I had a lot of ideas started in the Gherkin2.ps1 file, but, my
thoughts were getting jumbled. At the same time, I had some very much
working code and I just didn't want to delete it. But I needed a fresh
start.

So, here comes the Gherkin3.ps1 file. My approach is thus: following
along in _The Cucumber Book_, use an incremental approach to build out
the new Gherkin runner. The book leads a new user through setting up an
initial project, adding your first feature, etc. What better way to
approach writing the test runner?

So, I started simple with, I have no folders/files. And Invoke-Gherkin3
returns some helpful text to get you started. Then I started with the
`-Init` parameter, which creates the basic test folder structure for
you. And lastly I ended tonight with starting to write the code that
would find the feature files (actually, this code was already written in
Gherkin2.ps1 and was tested and known to be working). I just need to
figure out how to use the new Test Drive so I can get the tests to
successfully pass for this last bit.

I think this approach will be much more successful in allowing me to
incrementally build up the functionality. However, I'm also quickly
coming to the point where I need to know how Pester v5 screen display
works, as well as state management.

Because the primary test style in Pester is RSpec, much of the
"framework" is written specifically for that, and not for Gherkin--i.e.
it seems to rigid. Can it be improved to be a bit more flexible? I'm not
sure.