@wip Refactored getting support scripts -- again
Following after what Ruby's Cucumber does, and it makes sense, traverse
the file system under the present directory looking for all PowerShell
files. Ruby Cucmuber actually oes this several times--like, every time
the function that gets the files is called--but I only do it once.
Anyway, once we have this list of files, we filter the list to choose
what we need, e.g. support files vs. step definitions.

The benefit of this change is twofold: traversing the filesystem once;
and maknig the code much simpler for retrieving the support and step
definition files. This code will also translate to simpler code for
retrieving the feature files, too.