@wip Refactor getting feature files
Made the code simpler for getting feature files. Instead, we get
"potential" feature files. I'm not quite sure why they are "potential",
though this is the term Ruby Cucumber uses. So for now, I'm going to use
it.

But basically, we read "FeaturePathSpecs", which have the following
format:

   [DIR|FILE[:LINE]*]*

The LINE part specifies a line within a feature file that should be
executed, typically denoting the line upon which a scenario statement
occurs.

Also, the new Get-PotentialFeatureFile function also reads @Rerun.txt
files (actually, any file passed to Invoke-Gherkin in the format
@filename.ext is considered a rerun file), extracting "FeaturePathSpecs"
from the file. Each "FeaturePathSpec" will be on its own line.