@wip Adds ConvertTo-FileSpec
Cucumber allows you to specify paths, such as 'mydir:4',
'myfeature.feature:10', or even 'my_other_feature:11:21'. The `:x`
suffixes indicate lines within feature files to be run. For the case of
directories, a "line suffix" indicates that that line within each
feature located underneath that directory should be run. I see limited
utility in specifying line numbers at the directory level. It's highly
unlikely all features under a directory will have a scenario beginning
on the exact same line. So my implementation only parses line numbers
for specified feature files. This also means I won't be adding a -Line
parameter (to match Ruby Cucumber's --lines parameter) since you can
easily specify line numbers as part of the file path.

The CovertTo-FileSpec function converts 'my.feature:12:18' into an object
consisting of the filename portion and the lines to be run within the
feature file. As alluded to above, line numbers typically denote the
line on which a Scenario is defined in a feature file. The specified
lines, then, indicate that only those scenarios within the feature ought
to be exucted during the test run.

Never fear, support is coming for specifying the names of scenarios to
be executed, too. Really, this "filespec" is useful because it's a
compact format that can be written to the Rerun file, and then parsed by
Cucumber from the Rerun file for running failed scenarios within
features.