@wip Adds ability to obtain i18n translations
Adds the ability to list the supported Gherkin languages. Also adds the
ability to get all the translated keywords for a given language, or the
translations for a keyword for a given language.

Ruby Gherkin allows for the step definition keywords to be localized,
too. Not sure if that's possible with PowerShell. I'm thinking that,
depending on the language of a feature file to be executed, additional
aliases could be exported, if requried. Failing that, aliases could be
exported upfront for all translations?