Clean up Import-GherkinFeature
Cleaned up Import-GherkinFeature a bit in preparation for bigger
changes. Included in this changes is the Pester state object is a
mandatory parameter for Import-GherkinFeature. (I don't like magic
parameters that are sometimes passed, sometimes not, but still in scope.
If you have the parameter use it. Otherwise, make it a Script level
variable. I prefer the parameter.)

Making that parameter mandatory made some tests failed. I needed a way
to create and inject a Pester state object for the tests. So it seemed
like creating a New-GherkinPesterState function served that purpose,
whil also cleaning up Invoke-Gherkin a bit. Besides, there are changes
coming to that hunk of code anyway, so this will keep things nice and
neat and tidy.