Put report strings and theme in Gherkin.psd1
This commit takes the ReportTheme defined in Output.ps1 and places it as
a property in the object being created in Gherkin.psd1. Similarly, the
"top-level" properties in the Data object in Gherkin.psd1 were placed as
properties on a new object assigned to the top-level property
ReportStrings.

In this way, all information for formatting of the test console output
is in the Gherkin.psd1 file.

These changes are being made to support upcoming output changes. The
output changes will result in streamlined reporting to the console,
which will lead to perceived faster test runs (up to 50% in some cases)
and will be more inline with how Ruby Cucumber displays its output.