Adds NoMultiline parameter to Invoke-GherkinStep
Pivoting from working on Import-GherkinFeature, now I'm divivg deep and
working inside out. First up is qualifying some of the parameters to
Invoke-GherkinStep and adding the new NoMultiline switch parameter.

This parameter is similar to the --no-multiline option for Ruby Cucumber
and prevents output of step definition data tables and DocStrings. In an
upcoming commit, support will be added for displaying data tabales and
DocStrings in the console during the test run.