Invoke-GherkinStep improvements
This commit revamps Invoke-GherkinStep and adds several new
capabilities. It also simplified Invoke-GherkinScenario quite a bit by
taking advantage of a base-type inherited by both Gherkin Backgrounds
and Gherkin Scenarios and Scenario Outlines in Gherkin.dll v5.x,
allowing me to eliminate duplicate code, pushing that logic down into
Invoke-GherkinStep where it belongs.

The function was reorganized. The stop watch that times how long it
takes to execute the step definition, was moved very close to where the
step definition script block is actually executed. This more accurately
reports the time it takes to execue the step definition script block
itself, as opposed to the overhead execution time imposed by the runner
(which, I hope is recorded separately somewhere).

The logic was simplified for determining whether or not a step
definition should even be executed. All scenarios have a synthetic
'Result' property added to them when the feature files are imported and
parsed and the scenarios are processed. This made the simplification of
Invoke-GherkinScenario possible. This is much more efficient than
looping over all the steps for each scenario step to determine whether
or not the step should be executed; this should improve test run
execution times a bit. Once a step has a non-passing result, the code
checks whether the result is 'Inconclusive' (a/k/a in Gherkin,
undefined). If so, it attempts to determine whether or not the step that
would be executed exists or not. If not, the step is also marked as
'Inconclusive'. Otherwise, it is skipped. This is in keeping with how
Ruby Cucumber processes scenario steps. If a previous scenario step is
failed, then all future steps are skipped, regardless of whether or not
they may be undefined.

If no step definition is found for a scenario step, a new error record
is created using the new New-UndefinedStepErrorRecord function. If the
current scenario result is not 'Passed' or 'Inconclusive', then a new
error record is created using the New-StepSkippedErrorRecord function or
New-UndefinedStepErrorRecord function, as appropriate. If the step has
failed, the New-StepFailedErrorRecord function is used to capture
relevant information about the failure (i.e. stack trace, source
location, etc.). This new  function neatly encapsulates a lot of the
code that was previously just hanging out at the bottom of
Invoke-GherkinStep. Lastly, if the scenario is currently passing and
there is a step definition, but the new Set-StepPending function is
called, then that function calls New-StepPendingErrorRecord function to
return an error record for an undefined step.

There used to be code that would update the test result stack trace in
the event of a non-passing step. That's all been moved into the new
ConvertTo-GherkinStepResult function. There was sufficiently different
behavior from the ConvertTo-PesterResult function that it just makes
sense to not try to reuse it and instead create one specific for
Gherkin. Likewise, there was sufficient differences in displaying the
stack trace between Pester RSpec and Gherkin such that
ConvertTo-FailureLines wasn't really cutting it, and so a new
Format-FailedStepResult function was created. Currently, there is a lot
of commented out code, copied from ConvertTo-FailureLines, that resides
at the bottom of the function. This may be needed to show more detailed
stack traces--but if it is determined not to be needed, then that
commented out code will be removed prior to the PR implementing these
changes being merged into master.

Lastly, there was a check: `if ($Pester -and $Visible)` that is actually
erroneous. first, `$Pester` should always exist. Secondly, if `$Visible`
is not specified, it would mess up the step counts in the final test
report. I'm not sure that was the intent of the `-Visible` parameter. In
any event, the `-Visible` parameter was always used with
`Write-PesterResult` (now called Write-GherkinStepResult), so the
parameter was simply removed altogether. Again, there will be sufficient
differences in how step results are displayed in the console as compared
to Pester RSpec-style tests, such that Write-PesterResult was replaced
with Write-GherkinStepResult for Gherkin-style tests. This new function
also takes a new parameter, `-NoMultiline` (as does Invoke-Gherkin,
Invoke-GherkinFeature, Invoke-GherkinScenario, and Invoke-GherkinStep,
to pass it down when given from the command line), which controls
whether or not step definition DataTables and DocStrings are displayed
in the console. (NOTE: This parameter is not used yet--that will be in
an upcoming commit when printing DocStrings and DataTables is added.)