Show step DocStrings and DataTables in console
This commit changes how steps are displayed in the console. First off,
the code was regorganized to make smaller, more descriptive functions.
This commit also adds support for printing DocStrings and DataTables. In
addition, the steps no longer display with [+] or [-] to denote their
status.  Instead, color is used. This is in keeping with how Cucumber
displays its output. This change was made in support of further changes
to how console output will be displayed for features, scenarios, and
scenario outlines in upcoming commits.

Lastly, a change was made in how failing, undefined, and pending step
error messages and stack traces are displayed in the console. One aim
was to keep the format consistent so as to aid in viewing the errors in
the console. But another consideration was that of the Visual Studio
Code problem matchers. In the past, it has been difficult to make
changes to the Gherkin runner output without breaking the problem
matchers. While this change, too, breaks existing problem matchers, in
the end, it actually provides a consistent way of defining the problem
matchers, and so this is an improvement. Here are the problem matchers
I'm using in my VS Code tasks.json file for Pester Gherkin:

```json
{
  // Your Task definition here...
  "problemMatcher": [
    {
      "owner": "powershell",
      "fileLocation": ["absolute"],
      "severity": "error",
      "pattern": [
        {
          "regexp": "^\\s*Error:\\s+(.*)$",
          "message": 1
        },
        {
          "regexp": "^\\s+at\\s+[^,]+,\\s*(.*?):line\\s+(\\d+)$",
          "file": 1,
          "line": 2
        }
      ]
    },
    {
      "owner": "powershell",
      "fileLocation": ["absolute"],
      "severity": "warning",
      "pattern": [
        {
          "regexp": "^\\s*TODO: \\(Pester::Pending\\)$",
          "message": 0
        },
        {
          "regexp": "^\\s*at\\s+[^,]+,\\s*(.*?):line\\s+(\\d+)$",
          "file": 1,
          "line": 2
        }
      ]
    },
    {
      "owner": "powershell",
      "fileLocation": ["absolute"],
      "severity": "warning",
      "pattern": [
        {
          "regexp": "^\\s*>>> No matching step definition found.$",
          "message": 0
        },
        {
          "regexp": "^\\s*at\\s+(.*?):line\\s+(\\d+)$",
          "file": 1,
          "line": 2
        }
      ]
    }
  ]
}
```