Complete support for NoMultiline and Expand
This commit adds the NoMultiline and Expand parameters to Invoke-Gherkin
and ensures that they are passed down to where they're needed, paving
the way for the final set of changes that will produce MUCH nice output
with respect to Feature backgrounds and Scenario Outlines on par with
Ruby Cucumber.

Also, the `autoformatOnSave` setting auto-formatted some things in this
file. :|