Improve Gherkin PesterState, more cleanup
This commit contains some more "cleanup" items before moving on to the
last part that constitutes the main goal of this PR: making the
display of Pester Gherkin test runs in the console on par with Ruby
Cucumber test runs. This has a few benefits: the output in the console
will look very similar to your feature files. This means that the
background steps won't be printed over and over and over and over again
for each scenario in a feature. They'll be displayed just once. You
can't imagine how much time this can potentially shave off of a test
run.

The Gherkin PesterState object was updated to include not only the names
of passed/failed scenarios and passed/failed count, but also the
pending/undefined secnarios and their counts, too. The code used to
retrieve the scenarios for each status was simplified. This was made
possible by using a TestGroup hint of 'Feature' rather than simply
'Script' and by ensuring the TestGroup's Describe and Context properties
had values (the feature name and scenario name, respectively).

The Invoke-GherkinFeature and Invoke-GherkinScenario functions had some
more reorganization performed. For features, all the code that "sets up"
the feature before actually enumerating over the scenarios to execute
them was moved closer to the scenario enumeration. In fact, that "setup"
code was moved to after all the tag filtering code. In addition, the
"feature teardown code" was moved to a finally block. Previously, the
line `Invoke-GherkinHook AfterEachFeature ...` was executed after we had
changed locations back to the location where the test run was started
from. This was sa bug, as the AfterEachFeature hook might have dependend
on the location set in Invoke-Gherkin. (This reorganization is important
because in a future commit, most of that tag filtering code will move to
further improve test run execution times.)

Likewise, a similar reorganization was made in Invoke-GherkinScenario.
"The World", as it's called in Ruby Cucumber, is setup at the very top
of the Try block where the secnario steps are enumerated and executed.
Likewise, "the World" is torn down in an added finally block. And the
ordering of teardown was changed to mirror the ordering of setup--like
bookends. Again, a future commit will change how scenarios and scenario
outlines are processed, and these changes help towards organizing the
code for these upcoming changes.

Finally, with the addition of PendingScenarios and UndefinedScenarios
properties on the Pester state object for Gherkin tests, one of the
exitsing unit tests failed (prior to this commit, scenarios with
undefined steps were considered "failing"). So that test was corrected
and two new tests were created to cover these additional PesterState
properites. In addition, 3 more tests were added for testing that the
proper step results are returned. For example, if you have a failing
step early in your scenario, and a later step is undefined, the
scenario's result is failed, but the undefined step would still have a
result of undefined. This is inline with how Ruby Cucumber represents
the results of scenario steps under similar circumstances.

All-in-all, these changes once again bring us a bit closer to being on
par with Ruby Cucumber and how it processes gherkin features and step
definitions.