Fixes PSv2 PSObject Unwrapping, PSv6 Group-Object Sorting
So, this commit fixes two issues that ended up being interrelated to one
another--somewhat by accident.

A few commits back, I lamented that PSv2 was losing the PSObject wrapper
over the Scenario and Background objects in Import-GherkinFeature when
passing them to Invoke-GherkinScenario (actually, the wrapper was lost
when Invoke-GherkinScenario passed them along to Invoke-GherkinStep). At
the time, the workaround was to check if the ETS property Result was
still attached to the object, and if not, then reattach it. This had the
knock-on effect that I couldn't make use of the Result ETS property on
the Scenario object in the Pester TestResults object to easily determine
the passed, failed, pending and undefined scenarios. And I think this
has been a long-standing issue in the code since before I touched it,
leading to somewhat complicated code to determine the statu of the
secnario a'la ETS ScriptProperties added to the PesterState object.

I found a way to simplify that code a great deal. The only problem is,
as of PSv6.1, Group-Object now sorts the groups on their property key(s)
to achieve n*log(n) performance rather than n*n performance. Quite a
significant improvement. Though, I wasn't sure if the ordering of teh
secnario results was important at all--who knows how the Results object
is being used once it's left Pester.

If only I could use that Result ETS property on the scenarios. Because
then I wouldn't have to do any grouping on the individual step results,
but could query the secnarios directly. It would be faster, and simpler,
and maintain the order of the scenarios. So I decided to find out why
the PSObject wrapper was being lost along the way for PSv2.

It turns out, unlike in PSv3 and later, when you specify a parameter
type directoly, such as `[Gherkin.Ast.Scenario]`, PSv2 coerces the type
back to the type specified (so long as there's a PSTypeName for it or PS
can convert it), thereby losing the PSObject ATS/ETS wrapper. I left the
types of the paramteres commented out in the code, more as a reminder.
Plus, in Pester v5, PSv2 support is being dropped and this whole issue
goes away, allowing us to use more "staically-typed" parameters for
type-safety.

Anyway, fixing that issue allowed the ETS Result property to flow all
the way through the test run so that it was available on the Scenario
objects which where added as a collection ETS property on the Feature
object, which is included in the final test results object. This allowed
me to push down all of those properties directly onto the
Pester.Gherkin.Result object returned when using the -PassThru
parameter, and use much simpler code to determine the result of any one
scenario.

Nice.

At this point, I'm ready for the last set of changes to better display
features, scenarios and backgrounds during the test run, bringing Pester
Gherkin nearly on par with Ruby Cucumber.