Process scenario outlines similarly to Ruby Cucumber
This commit builds on the work of @cgnuechtel via #1167 and PR #1168 and
brings the implementation more in line with how Ruby Cucumber processes
scenario outlines, especially with respect to how the scenario outline
example table scenarios are named and subsequently displayed in the
console.

This current commit stil prints all the scenario outline scenarios "long
form". This form is as if you had specified --expand with Ruby Cucumber,
or the -Expand parameter recently added to Pester Gherkin in a previous
commit.

In a future commit, the "compact form" will be implemented, which will
streamline the output, as well as have the knock on effect of reducing
the test run time because not as much printing to the console will
occur, which is a slow operation.

NOTE: I found a bug earlier today in the NUnit XML generation. While I
fixed some unit tests with respect to XPaths for how secnario outline
example scenarios are nested in the NUnit XML, I did not fix the
underlying issue with NUnit XML generation yet as reported on #1285.