Improves Gherkin console test run output
This commit improves Gherkin test run console output. The "theme" has
been updated to match the output used by the Ruby Cucumber pretty-print
formatter.

Among the improvements are:

* The default "margin" is 2 spaces. Each "level", e.g. scenario
outlines, scenarios, examples, etc., are indented by a multiple of this
margin. NOTE: Description text is not reformatted in any way--i.e. all
spacing and indentation in the feature file is preserved.
* The background is printed only once per feature, instead of once per
scenario. This leads to nicer looking output in the console. It's also
less repetitive in the console. And, due to these steps only being
displayed once, results in faster output (which leads to a perceived
increase in test run speed).
* Both exapanded and compact scenario outline example scenarios are
supported.