Adds support for unexpanded scenarios
This commit adds support for unexpanded scenarios. All this means in the
context of this commit is that the Scenario Outline example scenario
names are formatted such that, when they are printed in tabular form,
all the columns will line up nice and neat. However, this commit does
not add the support for printing out the compact secnario outline form.
That's coming up in a few commits. However, before adding support for
"pretty printing" of scenario outlines, pretty-printing of features and
scenarios themselves must be updated now that everything is in place.