Adds Write-GherkinReport
This commit adds a Gherkin-specific console test report summary which
matches the output of Ruby Cucumber detailing the number of Scenarios
for each scenario result, as well as the number of steps by result, and
the time it took for the test run.

At some point earlier in this PR, I had "pushed down" the creation of
the PassedScenarios, FailedScenarios, etc. properties from the
PesterState object to the test results object, forgetting that these
properties were used on the PesterState object for this console test
summary. But all was not in vain, because in moving this code around, I
found a very succinct way of obtaining this information. So even though
I have to move it back, the result is much cleaner.