New new runtime (#1326)
The original runtime executed the file twice and tried to match the code based on positions, this allows for code that is scattered among the tests to still run and have effects. This is the only upside. The downsides are that you need to make sure that block generators, test generators and test case generator functions run twice in the same way, because otherwise you'd get different amount of tests, and so on. In this PR I move from the file based execution to the generate and run execution (the first model).

This also isolates the discovery and run code better because now they don't need to share the same function and same scoping concerns.