Return a feature w/ its background and scenarios
This commit is mainly a cleanup commit. Instead of using the ,-operator,
use the @() syntax to ensure that the scenarios returned from
Import-GehrkinFeature are stored as an array. Also, instead of returning
a 3-tuple of Feature, Background, and Scenarios from
Import-GherkinFeature, the feature is more properly a parent of a
background and one or more scenario definitions (each of which are
either a scenario outline or a regular scenario). So just attach those
items to the Feature object and, if there are any scenarios, then return
the feature.

The rest were odds/ends, removing old comments, fixing spelling, etc.