Rework stack trace printing
Back when I first reworked console printing, I only emitted abberviated
stack traces, leaving a significant portion of code commented out. This
commit reinstates much of that code, but in a reorganized fashion.

This reorganization makes sense for the Gherkin runner. And this
reorganization allowed me to reduce the number of lines of code used for
step result printing by ~80%.

This change allows for the abbreviated stack traces that were being
printed before, as well as full stack traces (via the global
PesterDebugPreference_ShowFullErrors variable), useful for debugging the
Pester test framework itself.