Update StackTrace printing to work with PSv2
This commit refactors the previous stack trace rework code so that
printing stack traces is compatible with PowerShell v2.

In Gherkin.Core.ps1, there was code in Import-GherkinFeature which never
worked in PowerShell v2, but which did not throw a terminating error.
Instead, this was compensated for later. (I believe it was compensated
for when converting the result into a pester test result, where the
stack trace was massaged.)

Instead of using "compensating code" or having more than one way to do
the same thing (one for PSv2, one for everything else), I opted to
change the way information is stored and propagated in order to have the
information required to create nice, consistent stack traces.