Fix test report output for single scenario/step results
If a test run consisted of a single scenario, part of the regex used to match
the test output in order to colorize it failed to match because it was
looking for the word 'scenarios'. When only a single scenario is run,
the text 'scenarios' is replaced with the singular 'scenario'. So the
regex was updated so that the entire report line would be matched: e.g.

    Before this fix:

    1 undefined)

    After this fix:

    1 scenario (1 undefined)

A similar change was made for the code which reports the step results.