Fixes Convert-FailureLines for Core vs. Desktop PS (#1475)
* Fixes Convert-FailureLines for Core vs. Desktop PS

For the desktop version of Powershell, when getting the exception
message, there would be 3 lines: the exeception name and message, any
parameter name, and then the inner exception type (and first line
message):

<Exception Name>: <first message line>
Parameter: <param name>
<Inner Exception Name>: <first inner exn line>

But, in PowerShell Core (7.0?), the output looked like this:

<Exception name>: <first message line> (Parameter '<param name>')
<Inner Exception name>: <first inner exn line>

This commit detects whether the first message line matches a regex that
matches the parenthesized parameter name in the first message line, and
if so, splits the line into two. This "normalizes" the message lines
between PowerShell Desktop and PowerShell Core.