(GH-1655) Allow negation of Should -Invoke
Previously Should -Invoke would ignore a negative expectation i.e. Should -Not
-Invoke.  This commit updates the Should-Invoke and Should-InvokeInternal
functions to pass through the Negation and then change the error raising
behaviour.  In particular, using -Not should have the opposite effect of if
it was not negated. That is, if it raised an error, then the negated expectation
would not raise an error, and vice versa.

This commit also adds tests for using -Not.

Note that using -Not and -ExclusiveFilter is not a valid combination as it
does not behave in an obvious manner.