(GH-1817) Trigger a GH Action once published
This commit extends the release script to optionally trigger a GitHub Action
Workflow to update the Pester documentation site.  Note that this optional for
now until the required Github credentials are created and injected into the
release CI system (TeamCity?).  This commit also depends on the Docs GitHub
repository to have the correct Workflow in place.

See https://docs.github.com/rest/reference/actions#create-a-workflow-dispatch-event
for more information